// src/interfaces/graphQLInterface.tsx
// Generated by https://quicktype.io

export interface GraphQLPokemon {
  data: Data;
}

export interface Data {
  pokemon_v2_pokemon: PokemonV2Pokemon[];
}

export interface PokemonV2Pokemon {
  id: number;
  name: string;
  pokemon_v2_pokemontypes: PokemonV2Pokemontype[];
}

export interface PokemonV2Pokemontype {
  pokemon_v2_type: PokemonV2Type;
}

export interface PokemonV2Type {
  name: string;
}

export interface SimplePokemonGQL {
    id: number;
    name: string;
    pokemon_v2_pokemontypes: PokemonV2Pokemontype[];
    picture:  string;
    color?:  string;
}